+++ FadeBasic/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - `FadeSdk` errors use source maps to include file location in error messages
+- Parser reports error when an object initializer includes non assignment statements
+- Usage of global variables before they are declared reports a different error than simply an undefined symbol
 
 ## [0.0.44] - 2025-06-05